Better error handling for `make release`.

+++ makefile
@@ -1,17 +1,27 @@
-all:
+.PHONY: compile release
+
+all: compile
+compile:
 	./build.sh
-release: all
-	sed -i.bak 's/"version": "[^"]*"/"version": "$(VERSION)"/' selectize.jquery.json
-	sed -i.bak 's/"version": "[^"]*"/"version": "$(VERSION)"/' bower.json
+release:
+ifeq ($(strip $(version)),)
+	@echo "\033[31mERROR:\033[0;39m No version provided."
+	@echo "\033[1;30mmake release version=1.0.0\033[0;39m"
+else
+	sed -i.bak 's/"version": "[^"]*"/"version": "$(version)"/' selectize.jquery.json
+	sed -i.bak 's/"version": "[^"]*"/"version": "$(version)"/' bower.json
 	rm *.bak
+	make compile
 	cp selectize.js ../.selectize.js
 	git add .
-	git commit -a -m "Released $(VERSION)."
-	git tag v$(VERSION)
+	git commit -a -m "Released $(version)."
+	git tag v$(version)
 	git push origin master
 	git push origin --tags
 	git checkout gh-pages
 	mv -f ../.selectize.js js/selectize.js
-	git commit -a -m "Updated to $(VERSION)."
+	git commit -a -m "Updated selectize.js to latest version."
 	git push origin gh-pages
 	git checkout master
+	@echo "\033[32mv${version} released\033[0;39m"
+endif
